chore(tests-integration): rename modules
commit-id:6218b7d0
