+++ crates/tests-integration/src/lib.rs
@@ -1,4 +1,4 @@
+pub mod config_utils;
 pub mod flow_test_setup;
-pub mod integration_test_config_utils;
-pub mod integration_test_utils;
 pub mod state_reader;
+pub mod utils;
